backend/requirements.txt
uvicorn[standard]==0.22.0
sqlalchemy==1.4.52
psycopg2-binary==2.9.7
pydantic==1.10.12
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
joblib==1.3.2
pandas==2.2.2
xgboost==1.7.6
# lightgbm==3.3.5
lightgbm
python-magic==0.4.27
fastapi==0.95.2
boto3>=1.29,<1.35
mlflow==2.8.1
prometheus-client==0.17.1
evidently==0.3.2
email-validator==2.1.1
alembic